moved from rvn to rvn-cli on build command
