+++ rvn/rvn.go
@@ -2,16 +2,11 @@ package rvn
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"log"
-	"net/http"
-	"net/url"
 	"os"
 	"os/exec"
-	"strings"
 )
 
 // Types ======================================================================
@@ -347,189 +342,5 @@ func ReadTopo(src []byte) (Topo, error) {
 		fillInMissing(&topo.Switches[i].Host)
 	}
 
-	CheckRvnImages(topo)
-
 	return topo, nil
-}
-
-func CheckRvnImages(topo Topo) {
-	var images []string = make([]string, 0)
-	// FIXME: NO EXTENTIONS!!
-	// for each Node, get the image required
-	for i := 0; i < len(topo.Nodes); i++ {
-		var currentImg string = topo.Nodes[i].Host.Image
-		// no way to check existance
-		var exists bool = false
-		for j := 0; j < len(images); j++ {
-			if currentImg == images[j] {
-				exists = true
-				break
-			}
-		}
-		// we did not find image, add it to our image list
-		if exists != true {
-			images = append(images, currentImg)
-		}
-	}
-	// for each unique image, check that it exists, if not, download it
-	for i := 0; i < len(images); i++ {
-		// parse the uri reference (with golang url parser)
-		parsedURL, _ := url.Parse(images[i])
-
-		// NOTE: local images need to be prefixed with absolute path or ./ (accesible by root)
-		// NOTE: image name cannot contain qcow2! qcow2 is added by rvn in libvirt.go
-		remoteHost := parsedURL.Host
-
-		// if there is no host to contact remotely, look locally, see note above on naming
-		if remoteHost == "" {
-			// check if this is local image path, or deterlab path
-			splitPath := strings.Split(images[i], "/")
-			if len(splitPath) > 1 {
-				filePath := "/var/rvn/img/user/" + splitPath[len(splitPath)-1]
-				_, err := os.Stat(filePath)
-				// if there is an err, file does not exist, lets also check that the file
-				// exists at the pathway, if it does, copy file to /var/rvn/img
-				if err != nil {
-					_, err = os.Stat(images[i])
-					if err != nil {
-						log.Fatalln("Copy failed - make sure if relative paths accessable by root (~)")
-						log.Fatalln(err)
-					} else {
-						// no error, so image does exist, so lets copy from path to /var/rvn/img
-						log.Println("Attempting copy from: " + images[i] + " to: " + filePath)
-						err = CopyLocalFile(images[i], filePath)
-						if err != nil {
-							log.Fatalln(err)
-						}
-					}
-				}
-				// is only given by a name, so we can therefore assume it is a deterlab.net image
-			} else {
-				filePath := "/var/rvn/img/" + images[i]
-				_, err := os.Stat(filePath)
-				// if there is an err, file does not exist, download it
-				if err != nil {
-					// FIXME: shhh dont tell ryan about the extentions, need to rehost rvn qcows
-					remotePath := "https://mirror.deterlab.net/rvn/img/" + images[i] + ".qcow2"
-					log.Println("Attempting copy from: " + remotePath + " to: " + filePath)
-					var dl_err error = DownloadFile(filePath, remotePath)
-					// we tried to find the image on deterlab mirror, but could not, error
-					if dl_err != nil {
-						log.Fatalln(dl_err)
-					}
-				}
-			}
-			// if the host is remote, we should assume we are going off-world to get image
-		} else {
-			subPath, imageName, _ := ParseURL(parsedURL)
-			filePath := "/var/rvn/img/user/" + subPath
-			_, err := os.Stat(filePath + imageName)
-			// path to image does not exist, we will need to download it
-			if err != nil {
-				// first we create all the subdirectorys in the file path
-				cr_err := os.MkdirAll(filePath, 0755)
-				if cr_err != nil {
-					log.Fatalln(cr_err)
-				}
-				// now try to download the image to the correct lcoation
-				log.Println("Attempting copy from: " + parsedURL.String() + " to: " + filePath + imageName)
-				dl_err := DownloadURL(parsedURL, filePath, imageName)
-				if dl_err != nil {
-					log.Fatalln(dl_err)
-				}
-			}
-		}
-	}
-}
-
-// https://gist.github.com/elazarl/5507969
-func CopyLocalFile(src, dst string) error {
-	s, err := os.Open(src)
-	if err != nil {
-		return err
-	}
-	defer s.Close()
-	d, err := os.Create(dst)
-	if err != nil {
-		return err
-	}
-	if _, err := io.Copy(d, s); err != nil {
-		d.Close()
-		return err
-	}
-	return d.Close()
-}
-
-// return a path, which we will create a directory tree with path[0]/path[1]/.../path[n]/image
-func ParseURL(parsedURL *url.URL) (path string, image string, err error) {
-	// Path is easier to use than RawPath
-	remoteFullPath := parsedURL.Path
-	splitPath := strings.Split(remoteFullPath, "/")
-	// get the image name, dont let user specify qcow2
-	// when rvn goes beyond qcow2, need to use correct format
-	image = splitPath[len(splitPath)-1]
-	// get the scheme used
-	// create necessary variables
-	var userName string
-	var hostName string
-	// now to create a directory tree from the path, omit scheme and opaque
-	if parsedURL.Opaque != "" {
-		err = errors.New("Opaque URL not implemented")
-		return path, image, err
-	}
-	if parsedURL.User != nil {
-		userName = parsedURL.User.Username()
-		path = userName + "/"
-	}
-	if parsedURL.Host != "" {
-		hostName = parsedURL.Host
-		path = path + hostName + "/"
-	}
-	// ftp://user@host:/path will become user/host/path.../
-	pathMinusImage := strings.Join(splitPath[:len(splitPath)-1], "/")
-	path += pathMinusImage + "/"
-	return path, image, nil
-}
-
-func DownloadURL(parsedURL *url.URL, downloadPath string, imageName string) error {
-	URIScheme := parsedURL.Scheme
-	// if no scheme for downloading file is provided, default to https
-	// TODO: enforce HTTPS -- do not allow http, redirect
-	if URIScheme == "https" {
-		DownloadFile(downloadPath+imageName, parsedURL.String())
-	} else if URIScheme == "http" {
-		err := errors.New("http is not supported, please use https!")
-		return err
-	} else if URIScheme == "" {
-		DownloadFile(downloadPath+imageName, parsedURL.String())
-	} else {
-		err := errors.New(parsedURL.Scheme + " is not currently implemented!")
-		return err
-	}
-	return nil
-}
-
-// https://golangcode.com/download-a-file-from-a-url/
-func DownloadFile(filepath string, url string) error {
-	// Create the file
-	out, err := os.Create(filepath)
-	if err != nil {
-		return err
-	}
-	defer out.Close()
-
-	// Get the data
-	resp, err := http.Get(url)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	// Write the body to file
-	_, err = io.Copy(out, resp.Body)
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
